Remove glsks and net positions
