+++ flow-decomposition/src/test/java/com/powsybl/flow_decomposition/TestUtils.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2022, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.powsybl.flow_decomposition;
+
+import com.powsybl.iidm.import_.Importers;
+import com.powsybl.iidm.network.Country;
+import com.powsybl.iidm.network.Network;
+import org.junit.jupiter.api.Test;
+
+import java.nio.file.Paths;
+import java.util.Map;
+
+/**
+ * This class contains helper functions for tests.
+ *
+ * @author Hugo Schindler {@literal <hugo.schindler at rte-france.com>}
+ */
+class TestUtil {
+    static Network importNetwork(String networkResourcePath) {
+        String networkName = Paths.get(networkResourcePath).getFileName().toString();
+        return Importers.loadNetwork(networkName, AllocatedFlowTests.class.getResourceAsStream(networkResourcePath));
+    }
+}
